+++ src/lib.rs
@@ -39,6 +39,10 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+#[cfg(feature = "async")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async")))]
+pub mod async_api;
+
 mod bridge_support;
 mod callbacks;
 pub mod caption;